requirements.txt
altgraph==0.17
auto-py-to-exe==2.8.0
cffi==1.14.4
cycler==0.10.0
Eel==0.12.4
et-xmlfile==1.0.1
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
importlib-metadata==3.4.0
jdcal==1.4.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
openpyxl==3.0.6
pefile==2019.4.18
Pillow==8.1.0
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
six==1.15.0
whichcraft==0.6.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
